+++ extensions/g-cloud/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## [0.1.31] - {MERGE_DATE}
 - feature: Commands for each service.
 - feature: QuickSwitcher, quick switch between projects inside same service.
-- fix: Storage Service now retrive folders and sub-folders.
+- fix: Storage Service now retrieves folders and sub-folders.
 
 ## [0.1.30] - 2025-02-10
 - fix: SDK path now is not using Homebrew default path, allowing users to have full control.